Explorer plugin: open native context menu on Alt+Enter press
