+++ Plugins/Flow.Launcher.Plugin.Explorer/ContextMenu.cs
@@ -222,34 +222,7 @@ namespace Flow.Launcher.Plugin.Explorer
                             if (record.Type is ResultType.Volume)
                                 return false;
 
-                            var screenWithMouseCursor = System.Windows.Forms.Screen.FromPoint(System.Windows.Forms.Cursor.Position);
-                            var xOfScreenCenter = screenWithMouseCursor.WorkingArea.Left + screenWithMouseCursor.WorkingArea.Width / 2;
-                            var yOfScreenCenter = screenWithMouseCursor.WorkingArea.Top + screenWithMouseCursor.WorkingArea.Height / 2;
-                            var showPosition = new System.Drawing.Point(xOfScreenCenter, yOfScreenCenter);
-
-                            switch (record.Type)
-                            {
-                                case ResultType.File:
-                                {
-                                    var fileInfos = new FileInfo[]
-                                    {
-                                        new(record.FullPath)
-                                    };
-
-                                    new Peter.ShellContextMenu().ShowContextMenu(fileInfos, showPosition);
-                                    break;
-                                }
-                                case ResultType.Folder:
-                                {
-                                    var directoryInfos = new DirectoryInfo[]
-                                    {
-                                        new(record.FullPath)
-                                    };
-
-                                    new Peter.ShellContextMenu().ShowContextMenu(directoryInfos, showPosition);
-                                    break;
-                                }
-                            }
+                            ResultManager.ShowNativeContextMenu(record.FullPath, record.Type);
 
                             return false;
                         },
